ci: make all v1.1.0-rc lanes green
Three fixes to land the v1.1.0-rc CI matrix:

1. src/algorithms/grover.c
   printf("%6llu", uint64_t) is a format mismatch on Linux x86_64
   where uint64_t = `long unsigned int`, not `long long unsigned`.
   gcc 15 on ubuntu-22.04 + -Werror=format= treats this as fatal.
   Cast every uint64_t printf arg to (unsigned long long) -- 3 sites
   (search-space, target-state, found-state).

2. Bindings version sync to 1.1.0-rc
   10 binding manifest files were still pinned at 1.0.5:
       bindings/python/pyproject.toml
       bindings/rust/{moonlab,moonlab-sys,moonlab-tui}/Cargo.toml
       bindings/javascript/package.json (+ 5 sub-packages)
   The bindings_version_sync ctest reads VERSION.txt (1.1.0-rc) and
   compares against every manifest; all 10 now read 1.1.0-rc.
   Verified via tools/check_binding_versions.sh:
       All binding manifests in sync with VERSION.txt=1.1.0-rc.

3. Release workflow: switch from `make` to `cmake`
   .github/workflows/release.yml + docker/debian/release/Dockerfile
   were still using the legacy Makefile.  The Makefile path is broken
   since CMake became the source of truth: src/utils/manifest.c
   includes moonlab_build_info.h which only CMake's build/generated
   step produces, so `make` fails with:
       fatal error: moonlab_build_info.h: No such file or directory
   Switched all 6 build sites in release.yml (3 matrix + arm64-qemu +
   debian + python + rust + npm) plus the debian/release Dockerfile
   to the canonical `cmake -S . -B build && cmake --build build`
   pattern.  Also updated the README-style "From Source:" snippet in
   the release notes body to recommend cmake instead of bare `make`.

Local verification on macOS arm64:
   build/bindings_version_sync: 100% tests passed
   build/libquantumsim.dylib    : built clean

Pre-existing failure modes (unrelated to v1.1, not blocking):
   - macOS ASAN+UBSAN ctest "unit_libirrep_css" skipped (not regression)
   - test_manifest's libcuda probe behaviour on no-GPU runners

+++ src/algorithms/grover.c
@@ -275,13 +275,15 @@ void grover_print_result(const grover_result_t *result, const grover_config_t *c
     printf("║                                                           ║\n");
     printf("║  Search Configuration:                                    ║\n");
     printf("║    Qubits:              %6zu                            ║\n", config->num_qubits);
-    printf("║    Search Space:        %6llu (2^%zu)                   ║\n", 
-           1ULL << config->num_qubits, config->num_qubits);
-    printf("║    Target State:        %6llu                            ║\n", config->marked_state);
+    printf("║    Search Space:        %6llu (2^%zu)                   ║\n",
+           (unsigned long long)(1ULL << config->num_qubits), config->num_qubits);
+    printf("║    Target State:        %6llu                            ║\n",
+           (unsigned long long)config->marked_state);
     printf("║    Iterations:          %6zu                            ║\n", result->iterations_performed);
     printf("║                                                           ║\n");
     printf("║  Results:                                                 ║\n");
-    printf("║    Found State:         %6llu                            ║\n", result->found_state);
+    printf("║    Found State:         %6llu                            ║\n",
+           (unsigned long long)result->found_state);
     printf("║    Success:             %s                              ║\n",
            result->found_marked_state ? "✓ YES" : "✗ NO ");
     printf("║    Success Probability: %6.4f                          ║\n", result->success_probability);
